backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
bcrypt==3.2.2
cffi==2.0.0
click==8.3.2
Deprecated==1.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.136.0
greenlet==3.4.0
h11==0.16.0
httptools==0.7.1
idna==3.11
limits==5.8.0
Mako==1.3.11
MarkupSafe==3.0.3
packaging==26.1
passlib==1.7.4
pycparser==3.0
pydantic==2.13.1
pydantic-settings==2.13.1
pydantic_core==2.46.1
PyJWT==2.12.1
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
slowapi==0.1.9
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.2
